Add files via upload
Problem
Write a program which takes as input a CSV file listing people and the score they achieved in a test, and which outputs the top
scorers, those people with the highest marks, as well as what that mark was.
If there are two or more top scorers, the program should output all of them in alphabetical order. For instance, in the given
TestData.csv file, the output should be:
As this is an evaluation, your solution should implement parsing the CSV file from string, and not use standard libraries.
Notes
* Typically, we use the following programing languages Python and C#, but are open to receiving the code in other
programming languages.
* The application should include a way to input data via a plain-text file.
* The application must be able to run with the output directed to STDOUT.
* The application should output correct values for inputs other than those specified in the TestData.csv file.
* You should provide sufficient evidence that your application is complete by including an example using the
TestData.csv data.
* Write the data you read from TestData.csv into a database table (use any freely available database for local development).
The columns of the database table should be those in TestData.csv.
* The submission should be pushed to a public or private version-controlled repository (preferably GitHub) and the
relevant URL shared with us.
* NB You should not use standard CSV-parsing libraries – your solution should implement the CSV parsing from string
data.
* Separately, create a RESTful API with:
o A POST endpoint that will receive new scores and save them into the database table created in the program above.
o A GET endpoint that will retrieve from the database table the score of a specific person.
o A GET endpoint that will retrieve the top score(s). Remember the note in the problem statement about multiple top
scorers.
o Demonstrate how you would secure these two endpoints. It need not be through code, you may also elect to speak
about how you would solution for this requirement.
o If you had to host this API in the Cloud and create a user interface for it, what Cloud components would you use? Be
prepared to speak about why you elected those components.
Please include an explanation of your design choices and assumptions, along with your code and any instructions
180necessary to run the application.
Evaluation Criteria
* Rather than developing the most technically correct solution, we are interested in your thought process
and approach to deriving a solution to such a problem.
* This problem should be seen as an opportunity to showcase your knowledge and highlight those aspects of software
engineering which you believe to be important.
* So, while it is a small problem, we expect you to submit what you believe is a reasonably well polished application
that you would be able to:
o Run